+++ src/main/scala/util/RunUtils.scala
@@ -136,7 +136,7 @@ object IRLoading {
 
     val specification = IRLoading.loadSpecification(q.specFile, program, globals)
 
-    RunUtils.applyPCTracking(q, program, specification)
+    ir.transforms.PCTracking.applyPCTracking(q.pcTracking, program)
 
     IRContext(symbols, externalFunctions, globals, funcEntries, globalOffsets, specification, program)
   }
@@ -1135,48 +1135,6 @@ object RunUtils {
       topDownDSA = dsa.topDown.toMap,
       regionInjector = regionInjector
     )
-  }
-
-  def applyPCTracking(q: ILLoadingConfig, program: Program, spec: Specification): Unit = {
-
-    // convenience variable to hold all procedures with defined program counters.
-    val proceduresWithPCs = program.procedures.collect {
-      case p if p.address.isDefined => (p, p.address.get)
-    }
-
-    q.pcTracking match {
-      case PCTrackingOption.None =>
-        program.collect {
-          case x: Statement if x.label == Some("pc-tracking") =>
-            x.parent.statements.remove(x)
-
-          // note: direct jumps are from aslp and so do not have the pc-tracking tag
-          case x @ LocalAssign(Register("_PC", 64), _, _) =>
-            x.parent.statements.remove(x)
-        }
-        Logger.info(s"[!] Removed all PC-related statements")
-
-      case PCTrackingOption.Keep =>
-        Logger.info(s"[!] Removing PC-tracking assertion statements, keeping PC assignments")
-        program.collect { case x @ Assert(_, _, Some("pc-tracking")) =>
-          x.parent.statements.remove(x)
-        }
-      case PCTrackingOption.Assert =>
-        Logger.info(s"[!] Inserting PC-tracking requires/ensures")
-
-        proceduresWithPCs.foreach((proc, addr) => {
-          val pcVar = Register("_PC", 64)
-          val r30Var = Register("R30", 64)
-          val addrVar = BitVecLiteral(addr, 64)
-          val pcRequires = BinaryExpr(ir.EQ, pcVar, addrVar)
-          val pcEnsures = BinaryExpr(ir.EQ, pcVar, OldExpr(r30Var))
-
-          val name = proc.procName
-          proc.requiresExpr = pcRequires +: proc.requiresExpr
-          proc.ensuresExpr = pcEnsures +: proc.ensuresExpr
-          (name -> SubroutineSpec(name, requires = List(), ensures = List()))
-        })
-    }
   }
 }
